% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAUDE.R
\name{combineZStouffer}
\alias{combineZStouffer}
\title{Combines Z-scores using Stouffer's method}
\usage{
combineZStouffer(x)
}
\arguments{
\item{x}{a vector of Z-scores to be combined}
}
\value{
Returns a single Z-score.
}
\description{
This function takes a vector of Z-scores and combines them into a single Z-score using Stouffer's method.
}
\examples{
combineZStouffer(rnorm(10))
}
