% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAUDE.R
\name{findOverlappingElements}
\alias{findOverlappingElements}
\title{Find overlaps between guides and annotated elements}
\usage{
findOverlappingElements(guides, elements, guides.pos = "pos",
  guides.chr = "chr", elements.start = "st", elements.end = "en",
  elements.chr = "chr")
}
\arguments{
\item{guides}{a data.frame containing guide information including the guides genomic position in a column named guides.pos}

\item{elements}{a data.frame containing element information, as in a BED file, including the element's genomic start, end, and chromosome in columns named elements.start, elements.end, and elements.chr}

\item{guides.pos}{the name of the column in guides that contains the genomic position targeted by the guide (defaults to "pos")}

\item{guides.chr}{the name of the column in guides that contains the genomic chromosome targeted by the guide  (defaults to "chr")}

\item{elements.start}{the name of the column in elements that contains the start coordinate of the element (defaults to "st")}

\item{elements.end}{the name of the column in elements that contains the start coordinate of the element (defaults to "en")}

\item{elements.chr}{the name of the column in elements that contains the start coordinate of the element (defaults to "chr")}
}
\value{
Returns a new data.frame containing the intersection of elements and guides
}
\description{
Finds guides that overlap the elements of a BED-like data.frame (e.g. open chromatin regions) and returns a new data.frame containing those overlaps
}
\examples{
annotatedGuides = findOverlappingElements(myGuides, myElements)
}
