% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAUDE.R
\name{getNBGaussianLikelihood}
\alias{getNBGaussianLikelihood}
\title{Calculate the log likelihood of observed read counts}
\usage{
getNBGaussianLikelihood(x, mu, k, sigma = 1, nullModel, libFract)
}
\arguments{
\item{x}{a vector of guide counts per bin}

\item{mu}{the mean for the normal expression distribution}

\item{k}{the vector of total counts per bin}

\item{sigma}{for the normal expression distribution (defaults to 1)}

\item{nullModel}{the bin bounds for the null model (for no change in expression)}

\item{libFract}{the fraction of the unsorted library this guide comprises (e.g. from unsorted cells, or sequencing the vector)}
}
\value{
the log likelihood
}
\description{
Uses a normal distribution (N(mu,sigma)) to estimate how many reads are expected per bin under nullModel, and calculates the log likelihood under a negative binomial model.
}
\examples{
#usually not used directly
}
