% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAUDE.R
\name{getTilingElementwiseStats}
\alias{getTilingElementwiseStats}
\title{Find active elements by sliding window}
\usage{
getTilingElementwiseStats(experiments, normNBSummaries, tails = "both",
  location = "pos", chr = "chr", window = 500, minGuides = 5, ...)
}
\arguments{
\item{experiments}{a data.frame containing the headers that demarcate the screen ID, which are all also present in normNBSummaries}

\item{normNBSummaries}{data.frame of guide-level statistics as generated by findGuideHits()}

\item{tails}{whether to test for increased expression ("upper"), decreased ("lower"), or both ("both"); (defaults to "both")}

\item{location}{the name of the column in normNBSummaries containing the chromosomal location (defaults to "pos")}

\item{chr}{the name of the column in normNBSummaries containing the chromosome name (defaults to "chr")}

\item{window}{the window width in base pairs (defaults to 500)}

\item{minGuides}{the minimum number of guides in a window required for a test (defaults to 5)}

\item{...}{other parameters for getZScalesWithNTGuides}
}
\value{
a data.frame containing the statistics for all windows tested for activity
}
\description{
Tests guides for activity by considering a sliding window across the tested region and including all guides within the window for the test.
}
\examples{
allGuideLevelStats = findGuideHitsAllScreens(myScreens, allDataCounts, allBinStats)
elementLevelStatsTiling = getTilingElementwiseStats(myScreens, allGuideLevelStats, tails = "upper")
}
