% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAUDE.R
\name{getZScalesWithNTGuides}
\alias{getZScalesWithNTGuides}
\title{Calculate Z-score scaling factors using non-targeting guides}
\usage{
getZScalesWithNTGuides(ntData, uGuidesPerElement, mergeBy, ntSampleFold = 10)
}
\arguments{
\item{ntData}{data.frame containing the data for the non-targeting guides}

\item{uGuidesPerElement}{a unique vector of guide counts per element}

\item{mergeBy}{usually contains a data.frame containing the headers that demarcate the screen ID}

\item{ntSampleFold}{how many times to sample each non-targeting guide to make the Z score scale (defaults to 10)}
}
\value{
a data.frame containing a Z-score scaling factor, one for every number of guides and unique entry in mergeBy
}
\description{
Calculates scaling factors to calibrate  element-wise Z-scores by repeatedly calculating a set of "null" Z-scores by repeatedly sampling the given numbers of non-targeting guides per element.
}
\examples{
#not generally used directly
}
