% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAUDE.R
\name{calcFDRByExperiment}
\alias{calcFDRByExperiment}
\title{FDR correction per experiment}
\usage{
calcFDRByExperiment(experiments, x, tails)
}
\arguments{
\item{experiments}{a data.frame containing the headers that demarcate the screen ID, which are all also present in normNBSummaries}

\item{x}{data.frame of element-level statistics, including columns for every column in 'experiments' and a column named 'p.value'}

\item{tails}{whether to test for increased expression ("upper"), decreased ("lower"), or both ("both"); (defaults to "both")}
}
\value{
a numerical vector containing B-H Q values corrected separately for every experiment.
}
\description{
Perform Benjamini-Hochberg FDR correction of p-values within each experiment. The returned values correspond to Q values. This is run automatically within getTilingElementwiseStats and getElementwiseStats, and doesn't generally need to be used directly.
}
\examples{
allGuideLevelStats = findGuideHitsAllScreens(myScreens, allDataCounts, allBinStats)
elementLevelStatsTiling = getTilingElementwiseStats(myScreens, allGuideLevelStats, tails = "upper")
elementLevelStatsTiling$Q = calcFDRByExperiment(myScreens, elementLevelStatsTiling, "upper") #Q column should equal FDR column
}
