% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAUDE.R
\name{findGuideHitsAllScreens}
\alias{findGuideHitsAllScreens}
\title{Calculate guide-level stats for multiple experiments}
\usage{
findGuideHitsAllScreens(experiments, countDataFrame, binStats, ...)
}
\arguments{
\item{experiments}{a data.frame containing the headers that demarcate the screen ID, which are all also present in countDataFrame and binStats}

\item{countDataFrame}{a table containing one column for each bin (A-F) and another column for non-targeting guide (logical-"NT"), and unsorted abundance (NS), as well as columns corresponding to those in  'experiments'}

\item{binStats}{a bin model as created by makeBinModel, as well as columns corresponding to those in  'experiments'}

\item{...}{other parameters for findGuideHits}
}
\value{
guide-level stats for all experiments
}
\description{
Uses findGuideHits to find guide-level stats for each unique entry in 'experiments'.
}
\examples{
 allGuideLevelStats = findGuideHitsAllScreens(myScreens, allDataCounts, allBinStats)
}
